1、m_stat_conclusion表新增字段及Mapper修改 2、视力字段Float类型修改BigDecimal(精度丢失及计算问题)相关代码修改 3、StatReportService类中sonarqube检测报红问题解决

+++ myopia-business/business-core/stat-core/src/main/java/com/wupol/myopia/business/core/stat/domain/dto/BigScreenStatDataDTO.java
@@ -2,6 +2,8 @@ package com.wupol.myopia.business.core.stat.domain.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 /**
  * @Description
  * @Date 2021/3/18 15:06
@@ -49,11 +51,11 @@ public class BigScreenStatDataDTO {
     /**
      * 左眼视力
      */
-    private Float visionL;
+    private BigDecimal visionL;
     /**
      * 右眼视力
      */
-    private Float visionR;
+    private BigDecimal visionR;
     /**
      * 城市地区id
      */
